setuptools: Disable auto discovery
The latest release of setuptools 61.0 made a breaking change[1] and
because of this change 'pip install' fails with the following error.

~~
error: Multiple top-level packages discovered in a flat-layout:
['lib', 'spec', 'manifests', 'releasenotes'].
~~

Users that don't set 'packages', 'py_modules', or configuration'
are still likely to observe the auto-discovery behavior, which may
halt the build if the project contains multiple directories and/or
multiple Python files directly under the project root.

To disable auto discovery, one can do below in setup.py

~~
setuptools.setup(..,packages=[],..)
~~

or

~~
setuptools.setup(..,py_modules=[],..)
~~

[1] pypa/setuptools#3197

Signed-off-by: Su Yue <glass@fydeos.io>